teste/http-cats/requirements.txt
absl-py==0.7.1
aiofiles==0.4.0
aiohttp==3.5.4
alembic==1.0.11
APScheduler==3.6.0
asn1crypto==0.24.0
astor==0.8.0
async-generator==1.10
async-timeout==3.0.1
attrs==19.1.0
Automat==0.7.0
Babel==2.4.0
beautifulsoup4==4.7.1
blinker==1.4
boto3==1.9.146
botocore==1.12.187
catfish==1.4.4
certifi==2018.1.18
chardet==3.0.4
click==6.7
colorama==0.3.7
colorclass==2.2.0
coloredlogs==10.0
colorhash==1.0.2
command-not-found==0.3
ConfigArgParse==0.14.0
constantly==15.1.0
cryptography==2.1.4
cssselect==1.0.3
cupshelpers==1.0
cycler==0.10.0
decorator==4.4.0
defer==1.0.6
distro-info===0.18ubuntu0.18.04.1
docopt==0.6.2
docutils==0.14
fakeredis==1.0.3
fbmessenger==6.0.0
Flask==1.1.1
Flask-BabelEx==0.9.3
Flask-Cors==3.0.8
Flask-Gravatar==0.4.2
Flask-HTMLmin==1.3.2
Flask-Login==0.4.0
Flask-Mail==0.9.1
Flask-Migrate==2.1.1
Flask-Paranoid==0.2.0
Flask-Principal==0.4.0
Flask-Security==1.7.5
Flask-SQLAlchemy==2.1
Flask-WTF==0.14.2
future==0.17.1
gast==0.2.2
gevent==1.4.0
google-pasta==0.1.7
greenlet==0.4.15
grpcio==1.22.0
h5py==2.9.0
htmlmin==0.1.12
httplib2==0.11.3
httptools==0.0.13
humanfriendly==4.18
hyperlink==19.0.0
idna==2.6
idna-ssl==1.1.0
incremental==17.5.0
ipython-genutils==0.2.0
isodate==0.6.0
itsdangerous==0.24
Jinja2==2.10.1
jmespath==0.9.4
jsonpickle==1.1
jsonschema==2.6.0
jupyter-core==4.4.0
kafka-python==1.4.6
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
keyring==10.6.0
keyrings.alt==3.0
kiwisolver==1.1.0
language-selector==0.1
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
lightdm-gtk-greeter-settings==1.2.2
lxml==4.3.4
Mako==1.0.7
Markdown==3.1.1
MarkupSafe==1.0
matplotlib==3.0.3
mattermostwrapper==2.1
menulibre==2.2.0
mixpanel==4.4.0
mock==3.0.5
mugshot==0.4.0
multidict==4.5.2
netifaces==0.10.4
networkx==2.3
numpy==1.16.3
oauth==1.0.1
olefile==0.45.1
onboard==1.4.1
packaging==19.0
pandas==0.24.2
paramiko==2.0.0
parsel==1.5.1
passlib==1.7.1
pexpect==4.2.1
pika==1.0.1
Pillow==5.1.0
prompt-toolkit==2.0.9
protobuf==3.8.0
psutil==5.4.2
psycopg2==2.7.6.1
psycopg2-binary==2.8.2
pyasn1==0.4.2
pyasn1-modules==0.2.5
pycairo==1.16.2
pycrypto==2.6.1
pycups==1.9.73
pycurl==7.43.0.1
PyDispatcher==2.0.5
pydot==1.4.1
pygobject==3.26.1
PyHamcrest==1.9.0
pyinotify==0.9.6
PyJWT==1.7.1
pykwalify==1.7.0
pymongo==3.8.0
pyOpenSSL==17.5.0
pyparsing==2.4.0
PySocks==1.7.0
python-apt==1.6.4
python-crfsuite==0.9.6
python-dateutil==2.8.0
python-debian==0.1.32
python-debianbts==2.7.2
python-dotenv==0.10.3
python-editor==1.0.4
python-engineio==3.8.2.post1
python-magic==0.4.16
python-socketio==4.0.1
python-telegram-bot==11.1.0
pytz==2019.1
pyxdg==0.25
PyYAML==3.12
questionary==1.1.1
queuelib==1.5.0
rasa==1.1.5
rasa-sdk==1.1.0
rasa-x==0.19.5
redis==3.2.1
reportlab==3.4.0
requests==2.22.0
requests-toolbelt==0.9.1
requests-unixsocket==0.1.5
robobrowser==0.5.3
rocketchat-API==0.6.31
ruamel.yaml==0.15.94
s3transfer==0.2.1
sanic==19.3.1
Sanic-Cors==0.9.8
sanic-jwt==1.3.1
Sanic-Plugins-Framework==0.7.0
scikit-learn==0.20.2
scipy==1.2.1
Scrapy==1.6.0
SecretStorage==2.3.1
selenium==3.141.0
service-identity==18.1.0
sgt-launcher==0.2.4
simplejson==3.16.0
six==1.12.0
sklearn-crfsuite==0.3.6
slackclient==1.3.1
sortedcontainers==2.1.0
soupsieve==1.9.1
SQLAlchemy==1.3.5
sqlparse==0.2.4
sshtunnel==0.1.4
systemd-python==234
tabulate==0.8.3
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
terminaltables==3.1.0
tqdm==4.31.0
traitlets==4.3.2
twilio==6.26.3
Twisted==19.2.1
typing-extensions==3.7.4
tzlocal==1.5.1
ubuntu-drivers-common==0.0.0
ufw==0.36
ujson==1.35
unattended-upgrades==0.1
urllib3==1.22
uvloop==0.12.2
virtualenv==16.6.1
w3lib==1.20.0
wadllib==1.3.2
wcwidth==0.1.7
webexteamssdk==1.1.1
websocket-client==0.54.0
websockets==6.0
Werkzeug==0.15.4
wrapt==1.11.2
WTForms==2.1
xcffib==0.5.1
xkit==0.0.0
yarl==1.3.0
zope.interface==4.6.0
